feat: complete day 6, puzzle 2

+++ src/AdventOfCode2023.Day06/GameUtil.cs
@@ -2,20 +2,20 @@ namespace AdventOfCode2023.Day06;
 
 internal static class GameUtil
 {
-    internal static int WinOptions(
-        int time,
-        int distance)
+    internal static long WinOptions(
+        long time,
+        long distance)
     {
         var d = Math.Sqrt(time * time - 4 * distance);
 
-        var min = (int)Math.Ceiling((time - d) / 2d);
+        var min = (long)Math.Ceiling((time - d) / 2d);
 
         if (min * time - min * min <= distance)
         {
             min++;
         }
 
-        var max = (int)Math.Floor((time + d) / 2d);
+        var max = (long)Math.Floor((time + d) / 2d);
 
         if (max * time - max * max <= distance)
         {
